Added "YOU GOT A MOON!" overlay to image
The font used is MOON GET! (https://maxigamer.deviantart.com/art/Font-MOON-GET-A-Super-Mario-Odyssey-Typeface-732847350). The name of the achievement is also printed in a config-defined font. The correct moon image is also added to the output.

Fixed issue where images wouldn't be stretched to conform to a 16:9 ratio.

The bot will no longer generate multi moons.